requirements.txt
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-openai==0.2.9
langgraph==0.2.22
chromadb==0.5.3
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
python-dotenv==1.0.1
typing-extensions>=4.8.0

